HealthcareFoundation/TaskManager/backend/app/monitoring/health.py
# app/monitoring/health.py
import asyncio
import aiohttp
from sqlalchemy.orm import Session
from app.core.database import engine
from app.core.config import settings


class HealthChecker:
    @staticmethod
    async def check_database() -> bool:
        """Check database connectivity"""
        try:
            with engine.connect() as conn:
                conn.execute("SELECT 1")
            return True
        except Exception:
            return False
    
    @staticmethod
    async def check_redis() -> bool:
        """Check Redis connectivity (if using Redis)"""
        try:
            # Add Redis health check if using Redis
            return True
        except Exception:
            return False
    
    @staticmethod
    async def check_external_apis() -> Dict[str, bool]:
        """Check external API dependencies"""
        results = {}
        
        # Add any external API health checks here
        # Example:
        # try:
        #     async with aiohttp.ClientSession() as session:
        #         async with session.get("https://api.example.com/health") as response:
        #             results["external_api"] = response.status == 200
        # except Exception:
        #     results["external_api"] = False
        
        return results
    
    @classmethod
    async def comprehensive_health_check(cls) -> Dict[str, Any]:
        """Run all health checks"""
        return {
            "database": await cls.check_database(),
            "redis": await cls.check_redis(),
            "external_apis": await cls.check_external_apis(),
            "status": "healthy"  # Overall status logic can be added here
        }
